Ui-filter changes for pages
